Do not use "asyncio.run()" to run the "play" command
Use a direct function reference instead of "asyncio.run" to call the "play_songs()" function, which generates the error "A coroutine was expected, got None" when exiting. This is because, since this function relies on a synchronous infinite loop already running in a daemon thread, there is no point in calling it asynchronously.